+++ CHANGELOG.md
@@ -11,6 +11,6 @@
   * License banner to all *.go files
 * Added examples of library usage:
   * Error code wrapping [signer/](/examples/signer/)
-  * Error wrapping [error_wrapping/main.go](/examples/lucky_unlucky/main.go)
+  * Error wrapping [error_wrapping/main.go](/examples/coinflip/main.go)
   * Error re-wrapping [error_re_wrapping/main.go](/examples/error_re_wrapping/main.go)
   * New error creation [error_creating/main.go](/examples/error_creation/main.go)